+++ cfdm/data/__init__.py
@@ -26,10 +26,12 @@ from .gatheredarray import GatheredArray
 from .fullarray import FullArray
 from .h5netcdfarray import H5netcdfArray
 from .netcdfarray import NetCDFArray
+from .scipynetcdffilearray import ScipyNetcdfFileArray
 from .netcdf4array import NetCDF4Array
 from .netcdfindexer import netcdf_indexer
 from .numpyarray import NumpyArray
 from .pointtopologyarray import PointTopologyArray
+from .pyfivearray import PyfiveArray
 from .raggedcontiguousarray import RaggedContiguousArray
 from .raggedindexedarray import RaggedIndexedArray
 from .raggedindexedcontiguousarray import RaggedIndexedContiguousArray