+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -227,7 +227,9 @@ class NetCDFWrite(IOWrite):
 
         return array.flatten()
 
-    def _write_variable_attributes(self, parent, ncvar, extra=None, omit=()):
+    def _write_variable_attributes(
+        self, parent, ncvar, extra=None, omit=(), dtype=None
+    ):
         """Write variable attributes to the dataset.
 
         :Parameters:
@@ -239,6 +241,13 @@ class NetCDFWrite(IOWrite):
             extra: `dict`, optional
 
             omit: sequence of `str`, optional
+
+            dtype: optional
+                The data type of the variable in the netCDF file. May
+                be used to ensure that fill and missing values have
+                the correct data type.
+
+                .. versionadded:: (cfdm) NEXTVERSION
 
         :Returns:
 
@@ -272,7 +281,12 @@ class NetCDFWrite(IOWrite):
 
             data = self.implementation.get_data(parent, None)
             if data is not None:
-                dtype = g["datatype"].get(data.dtype, data.dtype)
+                if dtype is None:
+                    raise ValueError(
+                        "Must set dtype when attributes include _FillValue or "
+                        "missing_value"
+                    )
+
                 netcdf_attrs[attr] = np.array(netcdf_attrs[attr], dtype=dtype)
 
         skip_set_fill_value = False
@@ -3236,7 +3250,7 @@ class NetCDFWrite(IOWrite):
         # Write attributes to the dataset variable
         # ------------------------------------------------------------
         attributes = self._write_variable_attributes(
-            cfvar, ncvar, extra=extra, omit=omit
+            cfvar, ncvar, extra=extra, omit=omit, dtype=datatype
         )
 
         # ------------------------------------------------------------